reactive recommendations

+++ recommendation-service/src/main/java/org/psp/core/recommendation/service/RecommendationServiceImpl.java
@@ -1,6 +1,7 @@
 package org.psp.core.recommendation.service;
 
 import com.mongodb.DuplicateKeyException;
+import lombok.RequiredArgsConstructor;
 import org.psp.api.core.recommendation.Recommendation;
 import org.psp.api.core.recommendation.RecommendationService;
 import org.psp.api.exceptions.InvalidInputException;
@@ -10,11 +11,15 @@ import org.psp.util.http.ServiceUtil;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.RestController;
 import org.slf4j.Logger;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.logging.Level;
 
 @RestController
+@RequiredArgsConstructor
 public class RecommendationServiceImpl implements RecommendationService {
     private static final Logger LOG = LoggerFactory.getLogger(RecommendationServiceImpl.class);
 
@@ -22,45 +27,44 @@ public class RecommendationServiceImpl implements RecommendationService {
     private final RecommendationMapper mapper;
     private final ServiceUtil serviceUtil;
 
-    public RecommendationServiceImpl(RecommendationRepository repository, RecommendationMapper mapper, ServiceUtil serviceUtil) {
-        this.repository = repository;
-        this.mapper = mapper;
-        this.serviceUtil = serviceUtil;
-    }
-
     @Override
-    public Recommendation createRecommendation(Recommendation body) {
-        try {
-            RecommendationEntity entity = mapper.apiToEntity(body);
-            RecommendationEntity newEntity = repository.save(entity);
-
-            LOG.debug("createRecommendation: created a recommendation entity: {}/{}", body.getProductId(), body.getRecommendationId());
-            return mapper.entityToApi(newEntity);
-        } catch(DuplicateKeyException e) {
-            throw new InvalidInputException("Duplicate key, product Id: " + body.getProductId()
-                    + ", Recommendation id: " + body.getRecommendationId());
+    public Mono<Recommendation> createRecommendation(Recommendation body) {
+        if (body.getProductId() < 1) {
+            throw new InvalidInputException("Invalid productId: " + body.getProductId());
         }
+        RecommendationEntity entity = mapper.apiToEntity(body);
+        Mono<Recommendation> newEntity = repository.save(entity)
+                .log(LOG.getName(), Level.FINE)
+                .onErrorMap(DuplicateKeyException.class,
+                        ex -> new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Recommendation Id:" + body.getRecommendationId()))
+                .map(e -> mapper.entityToApi(e));
+        return newEntity;
     }
 
     @Override
-    public List<Recommendation> getRecommendations(int productId) {
+    public Flux<Recommendation> getRecommendations(int productId) {
         if (productId < 1) {
             throw new InvalidInputException("Invalid productId: " + productId);
         }
-
-        List<RecommendationEntity> entityList = repository.findByProductId(productId);
-        List<Recommendation> list = mapper.entityListToApiList(entityList);
-        list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
-
-        LOG.debug("getRecommendations: response size: {}", list.size());
-
-        return list;
+        LOG.info("Get recommendations for product with ID: {}", productId);
+        return repository.findByProductId(productId)
+                .log(LOG.getName(), Level.FINE)
+                .map(e -> mapper.entityToApi(e))
+                .map(e -> setServiceAddress(e));
     }
 
     @Override
-    public void deleteRecommendation(int productId) {
+    public Mono<Void> deleteRecommendations(int productId) {
+        if (productId < 1) {
+            throw new InvalidInputException("Invalid product: " + productId);
+        }
         LOG.debug("deleteRecommendations: tries to delete recommendations for the product with productId: {}",
                 productId);
-        repository.deleteAll(repository.findByProductId(productId));
+        return repository.deleteAll(repository.findByProductId(productId));
+    }
+
+    private Recommendation setServiceAddress(Recommendation recommendation) {
+        recommendation.setServiceAddress(serviceUtil.getServiceAddress());
+        return recommendation;
     }
 }
